parseInt in lines
I think it's better to convert lines to integer, because it can't be float.

+++ project.js
@@ -41,7 +41,7 @@ const deposit = () => {
 const getNumberOfLines = () => {
   while (true) {
     const lines = prompt("Enter the number of lines to bet on (1-3): ");
-    const numberOfLines = parseFloat(lines);
+    const numberOfLines = parseInt(lines);
 
     if (isNaN(numberOfLines) || numberOfLines <= 0 || numberOfLines > 3) {
       console.log("Invalid number of lines, try again.");
